Reduce model capability registry sprawl
